pyproject.toml
[tool.poetry]
name = "flaskr-tdd"
version = "0.1.0"
description = ""
authors = []
packages = [
  { include = "project" },
]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^1.1.2"
python-dotenv = "^0.19.0"
gunicorn = "^20.1.0"
Flask-SQLAlchemy = "^2.5.1"
psycopg2-binary = "^2.9.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
flake8 = "^3.9.2"
isort = "^5.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
